Se agregaron BussinessException

+++ src/main/java/com/neoris/turnosrotativos/repositories/ConceptoRepository.java
@@ -3,5 +3,5 @@ package com.neoris.turnosrotativos.repositories;
 import com.neoris.turnosrotativos.entities.Concepto;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ConceptoRepository extends CrudRepository<Concepto, Long> {
+public interface ConceptoRepository extends CrudRepository<Concepto, Integer> {
 }
